+++ massiv/src/Data/Massiv/Array/Manifest/Internal.hs
@@ -42,9 +42,6 @@ module Data.Massiv.Array.Manifest.Internal
   , sizeofMutableArray
   , iterateUntil
   , iterateUntilM
-  , unsafeSliceN
-  , unsafeOuterSliceN
-  , unsafeInnerSliceN
   ) where
 
 import Control.Exception (try)
@@ -142,7 +139,11 @@ instance Strategy M where
 instance Source M e where
   unsafeLinearIndex = mLinearIndex
   {-# INLINE unsafeLinearIndex #-}
-  unsafeLinearSlice ix sz arr = unsafeExtract ix sz (unsafeResize sz arr)
+
+  unsafeOuterSlice = unsafeOuterSliceN
+  {-# INLINE unsafeOuterSlice #-}
+
+  unsafeLinearSlice off sz arr = MArray (getComp arr) sz (unsafeLinearIndex arr . (+ off))
   {-# INLINE unsafeLinearSlice #-}
 
 
@@ -161,145 +162,19 @@ instance Resize M where
   unsafeResize !sz !arr = arr { mSize = sz }
   {-# INLINE unsafeResize #-}
 
-instance Index ix => Extract M ix e where
-  unsafeExtract !sIx !newSz !arr =
-    MArray (getComp arr) newSz $ \ i ->
-      unsafeIndex arr (liftIndex2 (+) (fromLinearIndex newSz i) sIx)
-  {-# INLINE unsafeExtract #-}
 
-
-unsafeSliceN ::
-     (MonadThrow m, Resize (R r), Extract r ix e, Index (Lower ix))
-  => Array r ix e
-  -> ix
-  -> Sz ix
-  -> Dim
-  -> m (Array (R r) (Lower ix) e)
-unsafeSliceN arr start cutSz dim = do
-  (_, newSz) <- pullOutSzM cutSz dim
-  return $ unsafeResize newSz (unsafeExtract start cutSz arr)
-{-# INLINE unsafeSliceN #-}
-
-instance Slice M Ix1 e where
-  unsafeSlice arr i _ _ = pure (unsafeLinearIndex arr i)
-  {-# INLINE unsafeSlice #-}
-
-instance Slice M Ix2 e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance {-# OVERLAPPING #-} Slice M Ix3 e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance HighIxN n => Slice M (IxN n) e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance Slice M Ix2T e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance Slice M Ix3T e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance Slice M Ix4T e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
-
-instance Slice M Ix5T e where
-  unsafeSlice = unsafeSliceN
-  {-# INLINE unsafeSlice #-}
 
 
 unsafeOuterSliceN ::
      forall r ix e. (Source r e, Index ix, Index (Lower ix))
   => Array r ix e
+  -> Sz (Lower ix)
   -> Int
   -> Array M (Lower ix) e
-unsafeOuterSliceN !arr !i =
-  MArray (getComp arr) (snd (unconsSz (size arr))) (unsafeLinearIndex arr . (+ kStart))
+unsafeOuterSliceN !arr szL !i = MArray (getComp arr) szL (unsafeLinearIndex arr . (+ kStart))
   where
     !kStart = toLinearIndex (size arr) (consDim i (zeroIndex :: Lower ix))
 {-# INLINE unsafeOuterSliceN #-}
-
-instance OuterSlice M Ix1 e where
-  unsafeOuterSlice !arr = unsafeIndex arr
-  {-# INLINE unsafeOuterSlice #-}
-
-instance OuterSlice M Ix2 e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance {-# OVERLAPPING #-} OuterSlice M Ix3 e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance HighIxN n => OuterSlice M (IxN n) e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance OuterSlice M Ix2T e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance OuterSlice M Ix3T e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance OuterSlice M Ix4T e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-instance OuterSlice M Ix5T e where
-  unsafeOuterSlice = unsafeOuterSliceN
-  {-# INLINE unsafeOuterSlice #-}
-
-
-instance InnerSlice M Ix1 e where
-  unsafeInnerSlice !arr _ = unsafeIndex arr
-  {-# INLINE unsafeInnerSlice #-}
-
-unsafeInnerSliceN ::
-     forall r ix e. (Index ix, Index (Lower ix), Source r e)
-  => Array r ix e
-  -> (Sz (Lower ix), Sz Int)
-  -> Int
-  -> Array M (Lower ix) e
-unsafeInnerSliceN !arr (szL, m) !i =
-  MArray (getComp arr) szL (\k -> unsafeLinearIndex arr (k * unSz m + kStart))
-  where
-    !kStart = toLinearIndex (size arr) (snocDim (zeroIndex :: Lower ix) i)
-{-# INLINE unsafeInnerSliceN #-}
-
-instance InnerSlice M Ix2 e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance {-# OVERLAPPING #-} InnerSlice M Ix3 e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance HighIxN n => InnerSlice M (IxN n) e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance InnerSlice M Ix2T e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance InnerSlice M Ix3T e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance InnerSlice M Ix4T e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
-
-instance InnerSlice M Ix5T e where
-  unsafeInnerSlice = unsafeInnerSliceN
-  {-# INLINE unsafeInnerSlice #-}
 
 instance Index ix => Load M ix e where
   loadArrayM scheduler (MArray _ sz f) = splitLinearlyWith_ scheduler (totalElem sz) f