+++ massiv/src/Data/Massiv/Array/Ops/Slice.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 -- |
@@ -62,24 +63,23 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 --     ]
 --   ]
 -- >>> arr !> 2
--- Array M Seq (Sz (2 :. 4))
+-- Array U Seq (Sz (2 :. 4))
 --   [ [ (2,0,0), (2,0,1), (2,0,2), (2,0,3) ]
 --   , [ (2,1,0), (2,1,1), (2,1,2), (2,1,3) ]
 --   ]
 --
--- There is nothing wrong with chaining, mixing and matching slicing operators, or even using them
--- to index arrays:
+-- There is nothing wrong with chaining, mixing and matching slicing operators:
 --
--- >>> arr !> 2 !> 0 !> 3
+-- >>> arr !> 2 !> 0 ! 3
 -- (2,0,3)
--- >>> arr !> 2 <! 3 ! 0
+-- >>> evaluateM (arr !> 2 <! 3) 0
 -- (2,0,3)
--- >>> (arr !> 2 !> 0 !> 3) == (arr ! 2 :> 0 :. 3)
+-- >>> (arr !> 2 !> 0 ! 3) == (arr ! 2 :> 0 :. 3)
 -- True
 --
 --
 -- @since 0.1.0
-(!>) :: OuterSlice r ix e => Array r ix e -> Int -> Elt r ix e
+(!>) :: (Index ix, Index (Lower ix), Source r e) => Array r ix e -> Int -> Array r (Lower ix) e
 (!>) !arr !ix = either throw id (arr !?> ix)
 {-# INLINE (!>) #-}
 
@@ -88,12 +88,15 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 -- `Nothing` when index is out of bounds.
 --
 -- @since 0.1.0
-(!?>) :: (MonadThrow m, OuterSlice r ix e) => Array r ix e -> Int -> m (Elt r ix e)
-(!?>) !arr !i
-  | isSafeIndex sz i = pure $ unsafeOuterSlice arr i
-  | otherwise = throwM $ IndexOutOfBoundsException sz i
-  where
-    !sz = fst (unconsSz (size arr))
+(!?>) ::
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e)
+  => Array r ix e
+  -> Int
+  -> m (Array r (Lower ix) e)
+(!?>) !arr !i = do
+  let (k, szL) = unconsSz (size arr)
+  unless (isSafeIndex k i) $ throwM $ IndexOutOfBoundsException k i
+  pure $ unsafeOuterSlice arr szL i
 {-# INLINE (!?>) #-}
 
 
@@ -104,15 +107,19 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 --
 -- >>> import Data.Massiv.Array
 -- >>> arr = makeArrayR U Seq (Sz (3 :> 2 :. 4)) fromIx3
--- >>> arr !?> 2 ??> 0 ??> 3 :: Maybe Ix3T
+-- >>> arr !?> 2 ??> 0 ?? 3 :: Maybe Ix3T
 -- Just (2,0,3)
--- >>> arr !?> 2 ??> 0 ??> -1 :: Maybe Ix3T
+-- >>> arr !?> 2 ??> 0 ?? -1 :: Maybe Ix3T
 -- Nothing
 -- >>> arr !?> 2 ??> -10 ?? 1
 -- *** Exception: IndexOutOfBoundsException: -10 is not safe for (Sz1 2)
 --
 -- @since 0.1.0
-(??>) :: (MonadThrow m, OuterSlice r ix e) => m (Array r ix e) -> Int -> m (Elt r ix e)
+(??>) ::
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e)
+  => m (Array r ix e)
+  -> Int
+  -> m (Array r (Lower ix) e)
 (??>) marr !ix = marr >>= (!?> ix)
 {-# INLINE (??>) #-}
 
@@ -120,30 +127,26 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 -- | /O(1)/ - Safe slice from the inside
 --
 -- @since 0.1.0
-(<!?) :: (MonadThrow m, InnerSlice r ix e) => Array r ix e -> Int -> m (Elt r ix e)
-(<!?) !arr !i
-  | isSafeIndex m i = pure $ unsafeInnerSlice arr sz i
-  | otherwise = throwM $ IndexOutOfBoundsException m i
-  where
-    !sz@(_, m) = unsnocSz (size arr)
+(<!?) :: (MonadThrow m, Index ix, Source r e) => Array r ix e -> Int -> m (Array D (Lower ix) e)
+(<!?) !arr !i = do
+  let (szL, m) = unsnocSz (size arr)
+  unless (isSafeIndex m i) $ throwM $ IndexOutOfBoundsException m i
+  pure $ unsafeInnerSlice arr szL i
 {-# INLINE (<!?) #-}
 
 
 -- | /O(1)/ - Similarly to (`!>`) slice an array from an opposite direction.
 --
 -- @since 0.1.0
-(<!) :: InnerSlice r ix e => Array r ix e -> Int -> Elt r ix e
-(<!) !arr !ix =
-  case arr <!? ix of
-    Right res -> res
-    Left exc  -> throw exc
+(<!) :: (Index ix, Source r e) => Array r ix e -> Int -> Array D (Lower ix) e
+(<!) !arr !ix = throwEither (arr <!? ix)
 {-# INLINE (<!) #-}
 
 
 -- | /O(1)/ - Safe slicing continuation from the inside
 --
 -- @since 0.1.0
-(<??) :: (MonadThrow m, InnerSlice r ix e) => m (Array r ix e) -> Int -> m (Elt r ix e)
+(<??) :: (MonadThrow m, Index ix, Source r e) => m (Array r ix e) -> Int -> m (Array D (Lower ix) e)
 (<??) marr !ix = marr >>= (<!? ix)
 {-# INLINE (<??) #-}
 
@@ -151,7 +154,8 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 -- | /O(1)/ - Same as (`<!>`), but fails gracefully with a `Nothing`, instead of an error
 --
 -- @since 0.1.0
-(<!?>) :: (MonadThrow m, Slice r ix e) => Array r ix e -> (Dim, Int) -> m (Elt r ix e)
+(<!?>) ::
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e) => Array r ix e -> (Dim, Int) -> m (Array D (Lower ix) e)
 (<!?>) !arr (dim, i) = do
   (m, szl) <- pullOutSzM (size arr) dim
   unless (isSafeIndex m i) $ throwM $ IndexOutOfBoundsException m i
@@ -161,7 +165,12 @@ infixl 4 !>, !?>, ??>, <!, <!?, <??, <!>, <!?>, <??>
 
 
 internalInnerSlice ::
-     (MonadThrow m, Slice r ix e) => Dim -> Sz ix -> Array r ix e -> Int -> m (Elt r ix e)
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e)
+  => Dim
+  -> Sz ix
+  -> Array r ix e
+  -> Ix1
+  -> m (Array D (Lower ix) e)
 internalInnerSlice dim cutSz arr i = do
   start <- setDimM zeroIndex dim i
   unsafeSlice arr start cutSz dim
@@ -176,18 +185,19 @@ internalInnerSlice dim cutSz arr i = do
 -- index is out of bounds or dimensions is invalid.
 --
 -- @since 0.1.0
-(<!>) :: Slice r ix e => Array r ix e -> (Dim, Int) -> Elt r ix e
-(<!>) !arr !dix =
-  case arr <!?> dix of
-    Right res -> res
-    Left exc  -> throw exc
+(<!>) :: (Index ix, Index (Lower ix), Source r e) => Array r ix e -> (Dim, Int) -> Array D (Lower ix) e
+(<!>) !arr !dix = throwEither (arr <!?> dix)
 {-# INLINE (<!>) #-}
 
 
 -- | /O(1)/ - Safe slicing continuation from within.
 --
 -- @since 0.1.0
-(<??>) :: (MonadThrow m, Slice r ix e) => m (Array r ix e) -> (Dim, Int) -> m (Elt r ix e)
+(<??>) ::
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e)
+  => m (Array r ix e)
+  -> (Dim, Int)
+  -> m (Array D (Lower ix) e)
 (<??>) !marr !ix = marr >>= (<!?> ix)
 {-# INLINE (<??>) #-}
 
@@ -205,15 +215,11 @@ internalInnerSlice dim cutSz arr i = do
 --   [ 2 :. 0, 2 :. 1 ]
 --
 -- @since 0.5.4
-outerSlices :: OuterSlice r ix e => Array r ix e -> Array D Ix1 (Elt r ix e)
-outerSlices arr = makeArray Seq k (unsafeOuterSlice arr)
+outerSlices ::
+     (Index ix, Index (Lower ix), Source r e) => Array r ix e -> Array D Ix1 (Array r (Lower ix) e)
+outerSlices arr = makeArray (getComp arr) k (unsafeOuterSlice (setComp Seq arr) szL)
   where
-    (k, _) = unconsSz $ size arr
--- TODO: move setComp to Load
--- outerSlices arr = makeArray (getComp arr) k (unsafeOuterSlice arr')
---   where
---     arr' = setComp Seq arr
---     (k, _) = unconsSz $ size arr
+    (k, szL) = unconsSz $ size arr
 {-# INLINE outerSlices #-}
 
 
@@ -229,15 +235,10 @@ outerSlices arr = makeArray Seq k (unsafeOuterSlice arr)
 --   [ 0 :. 1, 1 :. 1, 2 :. 1 ]
 --
 -- @since 0.5.4
-innerSlices :: InnerSlice r ix e => Array r ix e -> Array D Ix1 (Elt r ix e)
-innerSlices arr = makeArray Seq k (unsafeInnerSlice arr sz)
+innerSlices :: (Index ix, Source r e) => Array r ix e -> Array D Ix1 (Array D (Lower ix) e)
+innerSlices arr = makeArray (getComp arr) k (unsafeInnerSlice (setComp Seq arr) szL)
   where
-    sz@(_, k) = unsnocSz $ size arr
--- TODO: move setComp to Load
--- innerSlices arr = makeArray (getComp arr) k (unsafeInnerSlice arr' sz)
---   where
---     arr' = setComp Seq arr
---     sz@(_, k) = unsnocSz $ size arr
+    (szL, k) = unsnocSz $ size arr
 {-# INLINE innerSlices #-}
 
 -- | Create a delayed array of slices from within. Checks dimension at compile time.
@@ -287,10 +288,10 @@ innerSlices arr = makeArray Seq k (unsafeInnerSlice arr sz)
 --
 -- @since 0.5.4
 withinSlices ::
-     (IsIndexDimension ix n, Slice r ix e)
+     (IsIndexDimension ix n, Index (Lower ix), Source r e)
   => Dimension n
   -> Array r ix e
-  -> Array D Ix1 (Elt r ix e)
+  -> Array D Ix1 (Array D (Lower ix) e)
 withinSlices dim = either throwImpossible id . withinSlicesM (fromDimension dim)
 {-# INLINE withinSlices #-}
 
@@ -301,7 +302,11 @@ withinSlices dim = either throwImpossible id . withinSlicesM (fromDimension dim)
 -- /__Throws Exceptions__/: `IndexDimensionException`
 --
 -- @since 0.5.4
-withinSlicesM :: (MonadThrow m, Slice r ix e) => Dim -> Array r ix e -> m (Array D Ix1 (Elt r ix e))
+withinSlicesM ::
+     (MonadThrow m, Index ix, Index (Lower ix), Source r e)
+  => Dim
+  -> Array r ix e
+  -> m (Array D Ix1 (Array D (Lower ix) e))
 withinSlicesM dim arr = do
   (k, szl) <- pullOutSzM (size arr) dim
   cutSz <- insertSzM szl dim oneSz