+++ massiv/CHANGELOG.md
@@ -12,6 +12,7 @@
 * Remove `ix` from `Resize`
 * Prevent `showsArrayPrec` from changing index type
 * Change function argument to monadic action for `unstablePartitionM` and `unsafeUnstablePartitionM`
+* Replace `snull` with a more generic `isNull`
 
 # 0.6.1